Time: 0 ms (100.00%), Space: 45.1 MB (53.81%) - LeetHub

+++ 0268-missing-number/0268-missing-number.java
@@ -1,11 +1,10 @@
 class Solution {
     public int missingNumber(int[] nums) {
-        long exp=((nums.length)*(nums.length+1))/2;
+        int exp=((nums.length)*(nums.length+1))/2;
         for(int i=0;i<nums.length;i++){
            exp-=nums[i];
         }
-        
-        return (int)exp;
+        return exp;
         
     }
 }
